requirements.txt
requests>=2.33.0
pdfminer.six>=20221105
beautifulsoup4>=4.12.0
lxml>=5.0.0
yt-dlp>=2024.12.0
python-dateutil>=2.9.0
tenacity>=9.0.0
rich>=13.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
